requirements.txt
absl-py==2.2.2
ale-py==0.11.0
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
grpcio==1.71.0
gymnasium==1.1.1
Jinja2==3.1.6
kiwisolver==1.4.8
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
protobuf==6.30.2
psutil==7.0.0
pygame==2.6.1
Pygments==2.19.1
pytest==8.4.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
rich==14.0.0
sb3_contrib==2.6.0
setuptools==80.3.0
six==1.17.0
stable_baselines3==2.6.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torch==2.7.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
PyYAML==6.0.1
